Created new router for inference endpoints
