+++ src/app.py
@@ -4,11 +4,13 @@ import uvicorn
 
 from configuration import inject_dependencies
 from routers.v1.user_router import create_user_router
+from routers.v1.inference_router import create_inference_router
 
 
 def create_app() -> FastAPI:
     app: FastAPI = FastAPI()
     app.include_router(create_user_router())
+    app.include_router(create_inference_router())
     return app
